requirements.txt
streamlit==1.31.1
pandas==2.2.0
qrcode==7.4.2
Pillow==10.2.0
openpyxl==3.1.2 